simplify the client.ts cancellation code
